requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-texttospeech==2.14.1
pydantic==2.6.1
pytest==7.4.4
httpx==0.26.0
python-multipart==0.0.6 
cloudinary==1.37.0 
aio-pika==9.3.0 